refactor: 💡 modified field relation module
This commit will modify field relation module
