+++ src/modules/maintenance/field-relation/entities/field-relation.entity.ts
@@ -1,26 +1,31 @@
 import { Entity, JoinColumn, ManyToOne } from 'typeorm';
 
-import { Field } from '../../maintenance/field/entities/field.entity';
+import { Field } from '../../field/entities/field.entity';
 import { HRISBaseEntity } from 'src/core/entities/base-entity';
 
 @Entity('fieldrelation', { schema: 'public' })
-export class FieldRelation extends HRISBaseEntity{
-
+export class FieldRelation extends HRISBaseEntity {
   static plural = 'fieldRelations';
 
+  /**
+   * Many To One Relationship: FieldRelation and Field Entities
+   */
   @ManyToOne(type => Field, field => field.parentFieldRelations, {
     primary: true,
     nullable: false,
     onDelete: 'CASCADE',
   })
-  @JoinColumn({ name: 'parentfield' })
-  parentField: Field | null;
+  @JoinColumn({ referencedColumnName: 'uid' })
+  parentField: Field;
 
+  /**
+   * Many To One Relationship: FieldRelation and Field Entities
+   */
   @ManyToOne(type => Field, field => field.childFieldRelations, {
     primary: true,
     nullable: false,
     onDelete: 'CASCADE',
   })
-  @JoinColumn({ name: 'child_field' })
-  childField: Field | null;
+  @JoinColumn({ referencedColumnName: 'uid' })
+  childField: Field;
 }
